+++ src/index.ts
@@ -92,6 +92,9 @@ import { omitIndexHtml } from './rules/omit-index-html';
 import { enforceIdCapitalization } from './rules/enforce-id-capitalization';
 import { noUnusedUseState } from './rules/no-unused-usestate';
 import { noUuidv4Base62AsKey } from './rules/no-uuidv4-base62-as-key';
+import enforceDynamicFileNaming from './rules/enforce-dynamic-file-naming';
+import { default as preferUseCallbackOverUseMemoForFunctions } from './rules/prefer-usecallback-over-usememo-for-functions';
+import { noMarginProperties } from './rules/no-margin-properties';
 
 module.exports = {
   meta: {
@@ -206,6 +209,9 @@ module.exports = {
         '@blumintinc/blumint/omit-index-html': 'error',
         '@blumintinc/blumint/enforce-id-capitalization': 'error',
         '@blumintinc/blumint/no-unused-usestate': 'error',
+        '@blumintinc/blumint/enforce-dynamic-file-naming': 'error',
+        '@blumintinc/blumint/prefer-usecallback-over-usememo-for-functions': 'error',
+        '@blumintinc/blumint/no-margin-properties': 'error',
       },
     },
   },
@@ -236,6 +242,7 @@ module.exports = {
     'no-unused-props': noUnusedProps,
     'no-useless-fragment': noUselessFragment,
     'no-uuidv4-base62-as-key': noUuidv4Base62AsKey,
+    'enforce-dynamic-file-naming': enforceDynamicFileNaming,
     'prefer-fragment-shorthand': preferFragmentShorthand,
     'prefer-type-over-interface': preferTypeOverInterface,
     'require-memo': requireMemo,
@@ -307,5 +314,7 @@ module.exports = {
     'omit-index-html': omitIndexHtml,
     'enforce-id-capitalization': enforceIdCapitalization,
     'no-unused-usestate': noUnusedUseState,
+    'prefer-usecallback-over-usememo-for-functions': preferUseCallbackOverUseMemoForFunctions,
+    'no-margin-properties': noMarginProperties,
   },
 };